refactor: extract rag_tracing.py from rag_core (C-3, #560)
Move 11 tracing/telemetry symbols out of rag_core.py into a new leaf
module rag_tracing.py: TRACE_SCHEMA_VERSION, REDACTED_LIST_PLACEHOLDER,
_StageTimer, _attach_trace_diagnostics, summarize_stage_attempt,
build_query_rewrite_trace, build_planner_trace, build_result_trace,
redact_trace, strip_internal_scores, percentile, rate.

rag_tracing imports only from rag_metadata_processing (normalize_regions /
normalize_page_span for strip_internal_scores) and stdlib — no rag_core
dependency. rag_core re-exports all 11 names; existing callers
(tests/test_fuzzy_retrieval.py, eval/scorers/alignment.py,
eval/scorers/citation.py) unchanged.

Removes unused stdlib imports `copy` and `math` from rag_core.py.
